network_{read,write}.c: explicit conversion to size_t
The value of len comes from recv() or send(), and is either -1 or 0 or
positive.  We have already dealt with the case of (len == -1), so it is safe to
cast it to size_t.

+++ libcperciva/network/network_read.c
@@ -65,7 +65,7 @@ callback_buf(void * cookie)
 	}
 
 	/* We processed some data. */
-	C->bufpos += len;
+	C->bufpos += (size_t)len;
 
 	/* Do we need to keep going? */
 	if (C->bufpos < C->minlen)
